modified:   cmd/find_me_job/main.go 	modified:   internal/models/application.go 	modified:   internal/repository/application_repository.go 	modified:   internal/repository/company_repository.go 	modified:   internal/repository/vacancy_repository.go 	modified:   internal/services/company_service.go 	modified:   internal/transport/company_handler.go 	new file:   internal/validators/validators.go

+++ internal/repository/vacancy_repository.go
@@ -8,7 +8,7 @@ import (
 type VacancyRepository interface {
 	Search(models.VacancyFilter) ([]models.Vacancy, error)
 	Create(*models.Vacancy) error
-	GetByCompanyId(uint64) ([]models.Vacancy, error)
+	GetByCompanyId(uint) ([]models.Vacancy, error)
 	IsVacancyExists(id uint) (bool, error)
 }
 
@@ -33,7 +33,7 @@ func (r *vacancyRepository) Search(filter models.VacancyFilter) ([]models.Vacanc
 	return vacancies, nil
 }
 
-func (r *vacancyRepository) GetByCompanyId(id uint64) ([]models.Vacancy, error) {
+func (r *vacancyRepository) GetByCompanyId(id uint) ([]models.Vacancy, error) {
 	var vacancies []models.Vacancy
 
 	if err := r.db.Where("company_id = ?", id).Find(&vacancies).Error; err != nil {
